fastapi-docker/requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
requests
datetime
pydantic
typing
google-generativeai
Pillow
python-multipart